move helper funtions
